Reorganize.  Simplify.  DRY.
Reorganizes the generator into a simple, single generator.  I found the "subgenerator" idea a bit off for "mvc", because mvc is not as much a sub generator as it is an option on the main generator (modelled off of some of the yeoman supported generators). For this reason I moved it to an option.
DRYed a lot of the redundant stuff from the two sub generators.  Also simplified some of the common operations.  I bumped this to v0.3.0 because there it is a pretty big refactor and there is a breaking difference in that it is now yo express --mvc instead of the sub generator syntax of yo express:mvc.  I'm thinking that a subgenerator would be better suited for adding models to the mvc version or perhaps installing SASS/LESS/Stylus scaffolding.